Split the trend downsampler out of usage reads
reads.py held a query and a pure transform, so the name only half-fit.
downsample (which reads nothing) and its window ranges move to usage/trends.py;
list_finished_calls_today stays, and reads.py is now only reads. The agent
history cap moves into the gateway that owns it — the dashboard keeps its own —
dropping the private cross-module import.
